move NetworksDropDown into separate component
